LayoutTests/inspector/editor/text-editor-highlight-api-expected.txt

Test editor highlight regexp API.


Before highlight

<div class="inner-container" tabindex="0">
<div class="webkit-line-content">foo();</div>
<div class="webkit-line-content">foo_1; foor; foo_;</div>
<div class="webkit-line-content">foo_</div>
<div class="webkit-line-content">global_foo</div>
<div class="webkit-line-content">global_foo2</div>
<div class="webkit-line-content">some_other_foo_X</div></div>

Running: testHighlightRegex

Add highlight for regexp /foo_./
Lines repainted: 2

<div class="inner-container" tabindex="0">
<div class="webkit-line-content">foo();</div>
<div class="webkit-line-content">foo_1; foor; foo_;<span class="some-css-class text-editor-overlay-highlight" style="left: -1px; width: 37px;"> </span><span class="some-css-class text-editor-overlay-highlight" style="left: 90px; width: 37px;"> </span></div>
<div class="webkit-line-content">foo_</div>
<div class="webkit-line-content">global_foo</div>
<div class="webkit-line-content">global_foo2</div>
<div class="webkit-line-content">some_other_foo_X<span class="some-css-class text-editor-overlay-highlight" style="left: 76px; width: 37px;"> </span></div></div>

Running: testHighlightRange

Highlighting range from (1, 13) to (3, 16)
Lines repainted: 3

<div class="inner-container" tabindex="0">
<div class="webkit-line-content">foo();</div>
<div class="webkit-line-content">foo_1; foor; foo_;<span class="some-css-class text-editor-overlay-highlight" style="left: -1px; width: 37px;"> </span><span class="some-css-class text-editor-overlay-highlight" style="left: 90px; width: 37px;"> </span><span class="some-other-class text-editor-overlay-highlight" style="left: 90px; width: 37px;"> </span></div>
<div class="webkit-line-content">foo_<span class="some-other-class text-editor-overlay-highlight" style="left: -1px; width: 30px;"> </span></div>
<div class="webkit-line-content">global_foo<span class="some-other-class text-editor-overlay-highlight" style="left: -1px; width: 72px;"> </span></div>
<div class="webkit-line-content">global_foo2</div>
<div class="webkit-line-content">some_other_foo_X<span class="some-css-class text-editor-overlay-highlight" style="left: 76px; width: 37px;"> </span></div></div>

Running: testRemoveRegexHighlight

Remove highlight for regexp /foo_./
Lines repainted: 2

<div class="inner-container" tabindex="0">
<div class="webkit-line-content">foo();</div>
<div class="webkit-line-content">foo_1; foor; foo_;<span class="some-other-class text-editor-overlay-highlight" style="left: 90px; width: 37px;"> </span></div>
<div class="webkit-line-content">foo_<span class="some-other-class text-editor-overlay-highlight" style="left: -1px; width: 30px;"> </span></div>
<div class="webkit-line-content">global_foo<span class="some-other-class text-editor-overlay-highlight" style="left: -1px; width: 72px;"> </span></div>
<div class="webkit-line-content">global_foo2</div>
<div class="webkit-line-content">some_other_foo_X</div></div>

Running: testRemoveRangeHighlight

Remove highlight for range (1, 13) to (3, 16)
Lines repainted: 3

<div class="inner-container" tabindex="0">
<div class="webkit-line-content">foo();</div>
<div class="webkit-line-content">foo_1; foor; foo_;</div>
<div class="webkit-line-content">foo_</div>
<div class="webkit-line-content">global_foo</div>
<div class="webkit-line-content">global_foo2</div>
<div class="webkit-line-content">some_other_foo_X</div></div>
